fixup! fixup! fixup! fixup! tests: add thread_float test app

+++ tests/thread_float/main.c
@@ -80,7 +80,7 @@ static void *thread2(void *arg)
     f = init;
 
     while (1) {
-        for (unsigned int i = 0; i < 100000ul; i++) {
+        for (unsigned long i = 0; i < 100000ul; i++) {
             f = f + 1.0 / f;
         }
         init += 1.0;
